@@ -6,83 +6,66 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "dev-company-build": "ng build --app company --output-path dist/dev/company",
     "prod-build": "npm run prod-build-en & npm run prod-build-fr",
-    "prod-build-en": "ng build -prod --output-path dist/en/ --env=prod-en",
-    "prod-build-fr": "ng build -prod --output-path dist/fr/ --env=prod-fr",
     "prod-build-allForms-en": "npm run prod-company-build-external-en & npm run prod-company-build-internal-en & npm run prod-dossier-build-external-en & npm run prod-dossier-build-internal-en & npm run prod-transaction-build-en & npm run prod-applicationInfo-build-en",
     "prod-company-build": "npm run prod-company-build-external & npm run prod-company-build-internal",
     "prod-company-build-external": "npm run prod-company-build-external-en & npm run prod-company-build-external-fr",
-    "prod-company-build-external-en": "ng build -prod --output-path dist/prod/company/external/en/ --env=prod-en --app company",
-    "prod-company-build-external-fr": "ng build -prod --output-path dist/prod/company/external/fr/ --env=prod-fr --app company",
+    "prod-company-build-external-en": "ng run company:build --configuration=production-en --outputPath=dist/prod/company/en",
+    "prod-company-build-external-fr": "ng run company:build --configuration=production-fr --outputPath=dist/prod/company/fr",
     "prod-company-build-internal": "npm run prod-company-build-internal-en & npm run prod-company-build-internal-fr",
-    "prod-company-build-internal-en": "ng build -prod --output-path dist/prod/company/internal/en/ --env=prod-en --app company_internal",
-    "prod-company-build-internal-fr": "ng build -prod --output-path dist/prod/company/internal/fr/ --env=prod-fr --app company_internal",
-    "prod-dossier-build": "npm run prod-dossier-build-external & npm run prod-dossier-build-internal",
-    "prod-dossier-build-external": "npm run prod-dossier-build-external-en & npm run prod-dossier-build-external-fr",
-    "prod-dossier-build-external-en": "ng build -prod --output-path dist/prod/dossier/external/en/ --env=prod-en --app dossier",
-    "prod-dossier-build-external-fr": "ng build -prod --output-path dist/prod/dossier/external/fr/ --env=prod-fr --app dossier",
-    "prod-dossier-build-internal": "npm run prod-dossier-build-internal-en & npm run prod-dossier-build-internal-fr",
-    "prod-dossier-build-internal-en": "ng build -prod --output-path dist/prod/dossier/internal/en/ --env=prod-en --app dossier_internal",
-    "prod-dossier-build-internal-fr": "ng build -prod --output-path dist/prod/dossier/internal/fr/ --env=prod-fr --app dossier_internal",
+    "prod-company-build-internal-en": "ng run company_internal:build --configuration=production-en --outputPath=dist/prod/company_internal/en",
+    "prod-company-build-internal-fr": "ng run company_internal:build --configuration=production-fr --outputPath=dist/prod/company_internal/fr",
     "prod-transaction-build": "npm run prod-transaction-build-en & npm run prod-transaction-build-fr",
-    "prod-transaction-build-en": "ng build -prod --output-path dist/prod/transaction/en/ --env=prod-en --app transaction",
-    "prod-transaction-build-fr": "ng build -prod --output-path dist/prod/transaction/fr/ --env=prod-fr --app transaction",
+    "prod-transaction-build-en": "ng run transaction:build --configuration=production-en --outputPath=dist/prod/transaction/en",
+    "prod-transaction-build-fr": "ng run transaction:build --configuration=production-fr --outputPath=dist/prod/transaction/fr",
     "prod-applicationInfo-build": "npm run prod-applicationInfo-build-en & npm run prod-applicationInfo-build-fr",
-    "prod-applicationInfo-build-en": "ng build -prod --output-path dist/prod/applicationInfo/en/ --env=prod-en --app applicationInfo",
-    "prod-applicationInfo-build-fr": "ng build -prod --output-path dist/prod/applicationInfo/fr/ --env=prod-fr --app applicationInfo",
-    "test": "ng test",
-    "lint": "ng lint  ",
-    "e2e": "ng e2e",
-    "ng xi18n": "ng xi18n",
+    "prod-applicationInfo-build-en": "ng run applicationInfo:build --configuration=production-en --outputPath=dist/prod/applicationInfo/en",
+    "prod-applicationInfo-build-fr": "ng run applicationInfo:build --configuration=production-fr --outputPath=dist/prod/applicationInfo/fr",
     "compodoc": "./node_modules/.bin/compodoc -p src/tsconfig.json",
     "docs": "npm run docs:build",
     "docs:build": "compodoc -p tsconfig.json -n companyFormTest -d docs --hideGenerator",
     "docs:serve": "npm run docs:build -- -s",
     "docs:watch": "npm run docs:build -- -s -w",
-    "serve app1": "ng serve --app app1",
-    "serve company": "ng serve --app company",
-    "serve company-internal": "ng serve --app company_internal",
-    "serve dossier": "ng serve --app dossier",
-    "serve dossier-internal": "ng serve --app dossier_internal",
-    "serve transaction": "ng serve --app transaction",
-    "serve applicationInfo": "ng serve --app applicationInfo"
+    "serve company": "ng run company:serve",
+    "serve company-internal": "ng run company_internal:serve",
+    "serve transaction": "ng run transaction:serve",
+    "serve applicationInfo": "ng run applicationInfo:serve"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.1.1",
-    "@angular/cli": "1.5.5",
-    "@angular/common": "5.1.1",
-    "@angular/compiler": "5.1.1",
-    "@angular/core": "^5.1.1",
-    "@angular/forms": "5.1.1",
-    "@angular/http": "5.1.1",
-    "@angular/platform-browser": "5.1.1",
-    "@angular/platform-browser-dynamic": "5.1.1",
-    "@angular/router": "5.1.1",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0",
-    "@ngx-translate/core": "^9.1.1",
-    "@ngx-translate/http-loader": "^2.0.1",
+    "@angular-devkit/build-angular": "^0.803.20",
+    "@angular/animations": "8.2.14",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "8.2.14",
+    "@angular/http": "7.2.15",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
+    "@ng-bootstrap/ng-bootstrap": "^5.1.4",
+    "@ngx-translate/core": "^11.0.1",
+    "@ngx-translate/http-loader": "^3.0.1",
     "@types/file-saver": "^1.3.0",
-    "angular2-focus": "^1.1.2",
+    "angular2-focus": "^1.1.4",
     "classlist.js": "^1.1.20150312",
     "core-js": "^2.5.3",
     "file-saver": "^1.3.3",
     "ng2-select": "^2.0.0",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.5.3",
     "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.20"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/core": "^0.5.4",
-    "@angular/cli": "1.5.5",
-    "@angular/compiler-cli": "5.1.1",
-    "@angular/language-service": "5.1.1",
+    "@angular/cli": "^8.3.20",
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/language-service": "8.2.14",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/lodash": "^4.14.97",
     "@types/node": "~6.0.60",
-    "codelyzer": "^4.1.0",
+    "codelyzer": "^5.2.0",
     "compodoc": "0.0.41",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
@@ -92,11 +75,11 @@
     "karma-coverage-istanbul-reporter": "^1.4.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "lodash": "^4.17.4",
+    "lodash": "^4.17.15",
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "~2.4.2",
+    "typescript": "~3.5.3",
     "xsd4js": "^1.0.2"
   }
 }
